Fix undersubscription scenario by avoiding redundant wakeup
This used to go unnoticed if the sense was global but no longer does if senses are toggled.

+++ test/test-thread_squad.cpp
@@ -31,7 +31,7 @@ TEST_CASE("thread_squad")
     std::mutex mutex;
     auto threadId_Count = std::unordered_map<std::thread::id, int>{ };
     auto threadIndex_Count = std::unordered_map<int, int>{ };
-	int count = 0;
+    int count = 0;
 
     auto params = sysmakeshift::thread_squad::params{
         /*.num_threads = */ numThreads
@@ -46,16 +46,16 @@ TEST_CASE("thread_squad")
         auto lock = std::unique_lock<std::mutex>(mutex);
         ++threadId_Count[std::this_thread::get_id()];
         ++threadIndex_Count[ctx.thread_index()];
-		++count;
+        ++count;
     };
 
     SECTION("single task")
     {
         sysmakeshift::thread_squad(params).run(action);
 #ifdef THREAD_PINNING_SUPPORTED
-	    CHECK(threadId_Count.size() == static_cast<std::size_t>(numActualThreads));
+        CHECK(threadId_Count.size() == static_cast<std::size_t>(numActualThreads));
 #endif // THREAD_PINNING_SUPPORTED
-    	CHECK(threadIndex_Count.size() == static_cast<std::size_t>(numActualThreads));
+        CHECK(threadIndex_Count.size() == static_cast<std::size_t>(numActualThreads));
     }
 
     SECTION("fixed number of tasks")
@@ -63,42 +63,41 @@ TEST_CASE("thread_squad")
         int numTasks = GENERATE(0, 1, 2, 5, 10, 20);
         CAPTURE(numTasks);
 
-	    auto threadSquad = sysmakeshift::thread_squad(params);
-    	for (int i = 0; i < numTasks; ++i)
-	    {
-    	    threadSquad.run(action);
-	    }
+        auto threadSquad = sysmakeshift::thread_squad(params);
+        for (int i = 0; i < numTasks; ++i)
+        {
+            threadSquad.run(action);
+        }
 
-	    if (numTasks != 0)
-    	{
+        if (numTasks != 0)
+        {
 #ifdef THREAD_PINNING_SUPPORTED
-	        CHECK(threadId_Count.size() == static_cast<std::size_t>(numActualThreads));
+            CHECK(threadId_Count.size() == static_cast<std::size_t>(numActualThreads));
 #endif // THREAD_PINNING_SUPPORTED
-    	    CHECK(threadIndex_Count.size() == static_cast<std::size_t>(numActualThreads));
-	    }
+            CHECK(threadIndex_Count.size() == static_cast<std::size_t>(numActualThreads));
+        }
 
 #ifdef THREAD_PINNING_SUPPORTED
-	    for (auto const& id_count : threadId_Count)
-    	{
-        	CHECK(id_count.second == numTasks);
-	    }
+        for (auto const& id_count : threadId_Count)
+        {
+            CHECK(id_count.second == numTasks);
+        }
 #endif // THREAD_PINNING_SUPPORTED
-    	for (auto const& index_count : threadIndex_Count)
-	    {
-    	    CHECK(index_count.second == numTasks);
-	    }
-	}
+        for (auto const& index_count : threadIndex_Count)
+        {
+            CHECK(index_count.second == numTasks);
+        }
+    }
 
     SECTION("fixed number of tasks")
     {
-	    auto threadSquad = sysmakeshift::thread_squad(params);
-		for (int i = 1; i <= int(numActualThreads); ++i)
-		{
-			CAPTURE(i);
-			threadSquad.run(action, i);
-		}
+        auto threadSquad = sysmakeshift::thread_squad(params);
+        for (int i = 1; i <= int(numActualThreads); ++i)
+        {
+            CAPTURE(i);
+            threadSquad.run(action, i);
+        }
 
-		CHECK(count == numActualThreads * (numActualThreads + 1) / 2);
-	}
+        CHECK(count == int(numActualThreads * (numActualThreads + 1) / 2));
+    }
 }
-
